+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,5 +1,6 @@
 package com.techcourse;
 
+import com.interface21.webmvc.servlet.DispatcherServlet;
 import com.interface21.webmvc.servlet.mvc.handler.adapter.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mvc.handler.mapping.HandlerMappingRegistry;
 import jakarta.servlet.ServletContext;
