+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,0 +1,21 @@
+package com.techcourse.controller;
+
+import com.interface21.context.stereotype.Controller;
+import com.interface21.web.bind.annotation.RequestMapping;
+import com.interface21.web.bind.annotation.RequestMethod;
+import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.servlet.View;
+import com.interface21.webmvc.servlet.view.JspView;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+@Controller
+public class ForwardController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView root(final HttpServletRequest request, final HttpServletResponse response) {
+        final View view = new JspView("/index.jsp");
+
+        return new ModelAndView(view);
+    }
+}
