+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/mapping/ManualHandlerMapping.java
@@ -1,15 +1,11 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.handler.mapping;
 
-import com.interface21.webmvc.servlet.mvc.handler.mapping.HandlerMapping;
-import com.techcourse.controller.*;
-import jakarta.servlet.http.HttpServletRequest;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import com.interface21.webmvc.servlet.mvc.handler.Controller;
-import com.interface21.webmvc.servlet.mvc.handler.ForwardController;
-
+import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class ManualHandlerMapping implements HandlerMapping {
 
@@ -19,8 +15,6 @@ public class ManualHandlerMapping implements HandlerMapping {
 
     @Override
     public void initialize() {
-        controllers.put("/", new ForwardController("/index.jsp"));
-
         log.info("Initialized Handler Mapping!");
         controllers.keySet()
                 .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
